refactor: folder: structure
