+++ src/.env
@@ -1,0 +1,1 @@
+DATABASE_URL= mongodb://127.0.0.1:27017/schoolNode
