+++ src/errors/notFound.error.ts
